Remove support for installing from Julia dashboards
Currently, whenever we make a change to this repo then the Julia
related files are generated and committed too.

This commit removes all the existing compiled Julia files,
prevents them from being added back and adds some context to the
README being clear about this.

We based this work on the dash-bootstra-components repo:
https://github.com/facultyai/dash-bootstrap-components/blob/c38c943af6f65f4bd03d2c2e131ed4c5009be452/.gitignore#L39-L45